Add files via upload
The following project is based on Business Analysis for a pharmaceutical company. The project contains different types of Analysis, such as:
1. Trend analysis Product wise year on year (sales by value) 
2. Region wise Product sales analysis (sales by value)
3. Highest Sales according to Product over all years. (sales by value)
4. Sales percentage Product wise (sales by value)
5. Top 10 countries sales-wise (sales by value)
6. Sales growth/decline product-wise b/w FY 15 & FY 16 (sales by value)
7. Sales growth/decline region wise b/w FY 15 & FY 16 (sales by value)
8. Pivot table with slicers region-wise (sales by value & qty)
9. Pivot table with timeline (sales by value & qty)
10. Categories the products into high-value & high-volume products. Then create a 4-quadrant chart for both value & volume (bubble chart )
The tool used for the analysis Ms-Excel. The project uses a variety of excel functions and formulas for data cleaning, sorting, arranging, and final analysis. Finally, by using charts, visualization in the form of charts has been created and presented in PowerPoint.  The details of the individual analysis can be found in the charts itself; the attached Excel file contains all the formulas and analyses in the Analysis Sheet.